+++ OpenDiscordHaxx/SocketServer.cs
@@ -13,9 +13,9 @@ namespace DiscordHaxx
             _server.AddWebSocketService<DashboardEndpoint>("/dashboard");
             _server.AddWebSocketService<BotEndpoint>("/bot");
             _server.AddWebSocketService<RaidBotEndpoint>("/bot/raid");
-            _server.AddWebSocketService<CheckerEndpoint>("/bot/checker");
-            _server.AddWebSocketService<CleanerEndpoint>("/bot/cleaner");
-            _server.AddWebSocketService<ReconEndpoint>("/bot/recon");
+            _server.AddWebSocketService<CheckerEndpoint>("/checker");
+            _server.AddWebSocketService<CleanerEndpoint>("/cleaner");
+            _server.AddWebSocketService<ReconEndpoint>("/recon");
             _server.Start();
         }