+++ OpenDiscordHaxx/Tools/Checker/Endpoint.cs
@@ -10,70 +10,20 @@ namespace DiscordHaxx
 {
     public class CheckerEndpoint : WebSocketBehavior
     {
+        private static Checker _checker;
+
+
         protected override void OnOpen()
         {
-            Task.Run(() =>
+            if (_checker == null || _checker.Finished)
             {
-                if (Server.Bots.Count == 0)
-                {
-                    Send(new CheckerErrorRequest("notokens"));
-
-                    return;
-                }
-
-                Send(new CheckerStartedRequest());
-
-                int valid = 0;
-                int invalid = 0;
-                int total = Server.Bots.Count;
-                foreach (var client in new List<DiscordClient>(Server.Bots))
-                {
-                    BotCheckedRequest req = new BotCheckedRequest()
-                                                 { Bot = BotInfo.FromClient(client) };
-                    req.Progress.Total = total;
-
-                    try
-                    {
-                        client.JoinGuild("a");
-                    }
-                    catch (DiscordHttpException e)
-                    {
-                        req.Valid = e.Code == DiscordError.UnknownInvite;
-                    }
-                    catch (JsonReaderException)
-                    {
-                        Send(new CheckerErrorRequest("ratelimit"));
-
-                        break;
-                    }
-
-
-                    if (req.Valid)
-                        valid++;
-                    else
-                    {
-                        Server.Bots.Remove(client);
-                        invalid++;
-                    }
-
-                    req.Progress.Valid = valid;
-                    req.Progress.Invalid = invalid;
-
-
-                    Send(req);
-                }
-
-                if (total > Server.Bots.Count)
-                {
-                    StringBuilder tokensToAdd = new StringBuilder();
-                    foreach (var bot in Server.Bots)
-                        tokensToAdd.AppendLine(bot.Token);
-
-                    File.WriteAllText("Tokens-checked.txt", tokensToAdd.ToString());
-                }
-
-                Send(new CheckerRequest(CheckerOpcode.Done));
-            });
+                _checker = new Checker();
+                _checker.StartAsync();
+            }
+            else
+                Send(JsonConvert.SerializeObject(new CheckerProgress() { Total = _checker.Total,
+                                                                         Valid = _checker.Valid,
+                                                                         Invalid = _checker.Invalid }));
         }
     }
 }