+++ Changelog.md
@@ -1,6 +1,9 @@
 # 0.0.2.1
 
 - Added support for setting custom thread amounts on attacks.
+- Added cleaner and recon tools.
+- Made only 1 checker able to run at a time, making it harder to 
+overload the tokens.
 - Added Threads to the dashboard's 'Ongoing attacks' bot list.
 - Added more input handling.
 - Fixed a CPU usage issue.